+++ script/core/folding.lua
@@ -66,7 +66,8 @@ local care = {
     ['repeat'] = function (source, text, results)
         local start  = source.start
         local finish = source.keyword[#source.keyword]
-        if text:sub(finish - #'until' + 1, finish) ~= 'until' then
+        -- must end with 'until'
+        if #source.keyword ~= 4 then
             return
         end
         local folding = {