+++ changelog.md
@@ -9,6 +9,7 @@
 * `FIX` [#1125](https://github.com/sumneko/lua-language-server/issues/1125)
 * `FIX` [#1131](https://github.com/sumneko/lua-language-server/issues/1131)
 * `FIX` [#1134](https://github.com/sumneko/lua-language-server/issues/1134)
+* `FIX` [#1141](https://github.com/sumneko/lua-language-server/issues/1141)
 
 ## 3.2.3
 `2022-5-16`