+++ changelog.md
@@ -11,6 +11,7 @@
 * `FIX` [#1134](https://github.com/sumneko/lua-language-server/issues/1134)
 * `FIX` [#1141](https://github.com/sumneko/lua-language-server/issues/1141)
 * `FIX` [#1144](https://github.com/sumneko/lua-language-server/issues/1144)
+* `FIX` [#1150](https://github.com/sumneko/lua-language-server/issues/1150)
 
 ## 3.2.3
 `2022-5-16`