+++ changelog.md
@@ -12,6 +12,7 @@
 * `FIX` [#1141](https://github.com/sumneko/lua-language-server/issues/1141)
 * `FIX` [#1144](https://github.com/sumneko/lua-language-server/issues/1144)
 * `FIX` [#1150](https://github.com/sumneko/lua-language-server/issues/1150)
+* `FIX` [#1155](https://github.com/sumneko/lua-language-server/issues/1155)
 
 ## 3.2.3
 `2022-5-16`